namespace xmlrpc cache... into the xmlrpc package
