+++ tests/unit/test_config.py
@@ -321,7 +321,7 @@ def test_configure(monkeypatch, settings, environment, other_settings):
             pretend.call("pyramid_retry"),
             pretend.call("pyramid_tm"),
             pretend.call("pyramid_services"),
-            pretend.call(".xmlrpc_cache"),
+            pretend.call(".legacy.api.xmlrpc.cache"),
             pretend.call("pyramid_rpc.xmlrpc"),
             pretend.call(".legacy.action_routing"),
             pretend.call(".domain"),
