+++ warehouse/config.py
@@ -173,7 +173,7 @@ def configure(settings=None):
     maybe_set(settings, "statuspage.url", "STATUSPAGE_URL")
     maybe_set(settings, "token.password.secret", "TOKEN_PASSWORD_SECRET")
     maybe_set(settings, "token.email.secret", "TOKEN_EMAIL_SECRET")
-    maybe_set(settings, "xmlrpc_cache.url", "REDIS_URL")
+    maybe_set(settings, "warehouse.xmlrpc.cache.url", "REDIS_URL")
     maybe_set(
         settings,
         "token.password.max_age",
@@ -329,7 +329,7 @@ def configure(settings=None):
     config.include("pyramid_services")
 
     # Register our XMLRPC cache
-    config.include(".xmlrpc_cache")
+    config.include(".legacy.api.xmlrpc.cache")
 
     # Register support for XMLRPC and override it's renderer to allow
     # specifying custom dumps arguments.
