+++ warehouse/legacy/api/xmlrpc/cache/derivers.py
@@ -1,0 +1,60 @@
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import json
+
+from urllib.parse import urlparse
+
+from pyramid.exceptions import ConfigurationError
+
+from warehouse.legacy.api.xmlrpc.cache import interfaces
+from warehouse.legacy.api.xmlrpc.cache import services
+
+
+def cached_return_view(view, info):
+    if info.options.get('xmlrpc_cache'):
+        tag = info.options.get('xmlrpc_cache_tag')
+        expires = info.options.get('xmlrpc_cache_expires', 86400)
+        arg_index = info.options.get('xmlrpc_cache_arg_index')
+        slice_obj = info.options.get(
+            'xmlrpc_cache_arg_index',
+            slice(None, None)
+        )
+
+        def wrapper_view(context, request):
+            try:
+                service = request.find_service(interfaces.IXMLRPCCache)
+            except ValueError:
+                return view(context, request)
+            try:
+                key = json.dumps(request.rpc_args[slice_obj])
+                _tag = None
+                if arg_index is not None:
+                    _tag = tag % (request.rpc_args[arg_index])
+                return service.fetch(
+                    view, (context, request), {}, key, _tag, expires
+                )
+            except interfaces.CacheError:
+                return view(context, request)
+            response = view(context, request)
+            return response
+        return wrapper_view
+    return view
+
+
+cached_return_view.options = [
+    'xmlrpc_cache',
+    'xmlrpc_cache_tag',
+    'xmlrpc_cache_expires',
+    'xmlrpc_cache_arg_index',
+    'xmlrpc_cache_slice_obj',
+]
